src/slices/request/controllers/deleteRequest/deleteRequestController.ts
/* eslint-disable no-unsafe-optional-chaining */
import {
  HttpRequest,
  HttpResponse,
  Validation,
  badRequest,
  ok,
} from "@/application/helpers";
import { Controller } from "@/application/infra/contracts";
import { DeleteRequest } from "@/slices/request/useCases";

export class DeleteRequestController extends Controller {
  constructor(
    private readonly validation: Validation,
    private readonly deleteRequest: DeleteRequest
  ) {
    super();
  }
  async execute(httpRequest: HttpRequest<any>): Promise<HttpResponse<any>> {
    const errors = this.validation.validate(httpRequest?.query);
    if (errors?.length > 0) {
      return badRequest(errors);
    }
    const requestDeleteed = await this.deleteRequest({
      fields: { ...httpRequest?.query, createdById: httpRequest?.userId },
      options: {},
    });
    return ok(requestDeleteed);
  }
}
